requirements.txt
streamlit
pandas
numpy
plotly
openai
python-dotenv
yfinance
reportlab
pydantic>=2.0